chore: remove old list and move tags utils
