+++ lib/experimental/Layouts/Utils/TagsList/index.stories.tsx
@@ -47,6 +47,13 @@ const meta: Meta = {
       },
     ],
   },
+  decorators: [
+    (Story) => (
+      <div className="w-72">
+        <Story />
+      </div>
+    ),
+  ],
 }
 
 export default meta
