+++ lib/experimental/Layouts/Utils/exports.tsx
@@ -1,4 +1,4 @@
-export * from "./BadgesList"
 export * from "./DetailsItem"
 export * from "./DetailsItemsList"
+export * from "./TagsList"
 export * from "./VerticalSeparator"
