+++ lib/experimental/Widgets/Content/exports.tsx
@@ -3,7 +3,6 @@ export * from "./EventsList"
 export * from "./IconText"
 export * from "./IconTextsList"
 export * from "./IndicatorsList"
-export * from "./Tag"
 export * from "./TasksList"
 export * from "./TimeStatus"
 export * from "./TwoColumnsList"
